feat: initialize signals sem

+++ server/src/args.rs
@@ -1,7 +1,7 @@
 use clap::{Arg, Command, crate_version};
 
 
-pub fn parse_args() -> i32 {
+pub fn parse_args() -> usize {
     let matches = Command::new("ht")
         .version(crate_version!())
         .about("A hash table that supports concurrent operations")
@@ -15,7 +15,7 @@ pub fn parse_args() -> i32 {
         .get_matches(); // parse env::args_os
 
     let ht_size = matches.get_one::<String>("hash_table_size").unwrap();
-    let ht_size: i32 = ht_size.parse().expect("provided hash table size is not a number!");
+    let ht_size: usize = ht_size.parse().expect("provided hash table size is not a number!");
 
     ht_size
 }
